Added synchronized version badging
To avoid the hassle of maintaining multiple version string, DendroPy
now maintains a single version string in dendropy.__init__.
If dendropy __init__ is called by itself, then it prints the version
information to stdout.
setup.py now takes advantage of this, and calls dendropy.__init__,
parses the stdout for the version, and uses that.
So, as long the as the version string in dendropy.__init__ is updated,
all will be well.