Fundamental reworking of how normalize_taxa block in Trees/Char Block works if no taxa_block is passed
Previously, a new TaxaBlock was created, assigned and populated. This resulted in extra bookeeping as then the old taxa block had to be removed from the list of taxa blocks in the dataset, and the new one added, and client code invariably did not do this. Now, if a taxa_block is not passed, the taxa-linked block (e.g. TreesBlock) will reuse its current taxa_block, clearing out existing entries and recreating them from scratch. This way,  *if* a reference to the TreeBlock's taxa_block already exists in the containing Dataset, nothing needs to be changed.