New test added to reveal a bug involving encoding splits on pruned trees.
If the first taxon is not in the tree when the splits are encoded,
the mask in NormalizedBitmaskDict will be wrong. Failure to have the
1-bit set will cause the mask to be complemented. Subsequent "anding"
of that mask to any split's bitmask will result in a 0 key in the
NormalizedBitmaskDict. Thus there will be only one key in the dict and
bitmask-dependent operations (such as clade summarization and tree-to-tree
distances) will be incorrect.