Fixing bug in contained_coalescent_tree function.
Fixing bug in dendropy.model.coalescent.contained_coalescent_tree.
`default_pop_size` was always being used for the root edge, even if a
different value was specified via an attribute of the root edge. This
commit fixes this by specifying the correct variable in the final call
to `coalesce_nodes`.