Add label lookup to improve nexusreader speed
When parsing the TAXLABELS block in NEXUS formatted trees, the parser
checks whether each token already exists in the taxon namespace with
get_taxon. However, get_taxon requires a linear scan of the taxon
namespace, which becomes increasingly expensive as the namespace grows
in size. In particular, if reading a tree into an empty namespace, each
taxon insertion will incur increasingly poor worst case search performance
as the taxon is not in the tree.

This change modifies _parse_taxlabels_statement: before the token
parsing loop, in a similar manner to how _lookup_label works, the taxon
namespace is enumerated and each label (with case sensitivity considered)
is added to a set.

As each token is parsed, it is checked in constant time against the set,
rather than requiring a linear scan of the taxon namespace with get_taxon.
If the token is not in the set, a new taxon is added as before with new_taxon.
If the token is in the set, it can be safely fetched with get_taxon.

This change significantly improves performance on large NEXUS trees. As part
of the COG-UK phylogenetics pipeline, this patch reduces the time to load
a tree containing 340,000 TAXLABELS from two hours to approximately five
minutes.

The change does not degrade the results of the current test suite.
Additionally I have provided two new tests and test data to ensure that
the new behaviour of _parse_taxlabels_statement is stable when attaching
a taxon namespace to a Tree and reading in multiple NEXUS files with
case sensitivity set to True and False.